fix: treat compiler warnings as errors and resolve all warnings
- Add -Wall -Wextra -Werror -Wpedantic -Wconversion -Wsign-conversion to CMakeLists.txt
- Fix unused parameter warnings with (void) casts in action.h and test mocks
- Fix sign-conversion warnings with static_cast in delayed_action.cpp, profile_manager.cpp, send_action.cpp
- Fix missing-braces warning in test_main.cpp
- Add dependabot.yml for automated GitHub Actions dependency updates (TASK-008)
- Add Mermaid diagram linting job to CI workflow (TASK-039)
- Add gcovr coverage report artifact to unit-tests job (TASK-014)
- Add test coverage badge to README.md

+++ lib/PedalLogic/src/delayed_action.cpp
@@ -28,7 +28,7 @@ void DelayedAction::execute()
 {
     if (! started)
     {
-        startTime = millis();
+        startTime = static_cast<uint32_t>(millis());
         started = true;
     }
     else
